+++ reikna/reduce.py
@@ -105,8 +105,14 @@ class Reduce(Computation):
 
         plan = plan_factory()
 
+        # Using algorithm cascading: sequential reduction, and then the parallel one.
+        # According to Brent's theorem, the optimal sequential size is O(log(n)).
+        # Setting it to the nearest power of 2 to simplify integer operations.
+        max_seq_size = helpers.bounding_power_of_2(helpers.log2(device_params.max_work_group_size))
+
         # FIXME: may fail if the user passes particularly sophisticated operation
-        max_reduce_power = device_params.max_work_group_size
+        max_group_size = device_params.max_work_group_size
+        max_reduce_power = max_group_size * max_seq_size
 
         if self._transpose_axes is None:
             # normal reduction
@@ -129,23 +135,30 @@ class Reduce(Computation):
         while part_size > 1:
 
             if part_size > max_reduce_power:
-                block_size = max_reduce_power
-                blocks_per_part = helpers.min_blocks(part_size, block_size)
+                seq_size = max_seq_size
+                block_size = max_group_size
+                blocks_per_part = helpers.min_blocks(part_size, block_size * seq_size)
                 cur_output = plan.temp_array(
                     (final_size, blocks_per_part), input_.dtype)
                 output_slices = (1, 1)
             else:
-                block_size = helpers.bounding_power_of_2(part_size)
+                if part_size > max_group_size:
+                    seq_size = helpers.min_blocks(part_size, max_group_size)
+                    block_size = max_group_size
+                else:
+                    seq_size = 1
+                    block_size = helpers.bounding_power_of_2(part_size)
                 blocks_per_part = 1
                 cur_output = output
                 output_slices = (len(cur_output.shape), 0)
 
-            if part_size % block_size != 0:
-                last_block_size = part_size % block_size
+            if part_size % (block_size * seq_size) != 0:
+                last_block_size = part_size % (block_size * seq_size)
             else:
-                last_block_size = block_size
+                last_block_size = block_size * seq_size
 
             render_kwds = dict(
+                seq_size=seq_size,
                 blocks_per_part=blocks_per_part,
                 last_block_size=last_block_size,
                 log2=helpers.log2, block_size=block_size,